Refactor sass pre-processing to be handled again by gulp
Even if jekyll has a built in sass compiler we still want gulp to
handle the processing. This allows us to import the npm installed
version of bootstrap for easy upgrading and maintenance of
bootstrap.

The way this was handled was to copy bootstrap source files into
the css/_sass directory which meant we had two copies of the
bootstrap source in our codebase, the npm installed one and the
copy in the _sass directory. This is obviously worse from a
maintenance perspective as it involves manual copying and also it
didn't work perfectly, it ignored glyphicons handling which was
explicitly not imported before the jekyll change but got imported
after the jekyll change overwriting the cdn glyphicon handling and
breaking icons in the app. This could have been fixed by explicitly
importing the _sass manual copy, but that would still leave two
copies of bootstrap in every tree. We could also then not install
bootstrap with npm, but that would make maintenance even harder as
you then have to manually go and fetch a new version of bootstrap.

Another option would be to declare a sass folder in the jekyll yaml
config file (which is removed with this commit). That sass folder
would have to be node_modules or top level folder which isn't
optimal.

Therefore the simplest and best solution was to have sass compiling
handled by gulp which is already run before jekyll is run in order
to generate the js scripts and sass compilation is faster than
scripts so we're not really losing anything.

Goodbye manual boostrap copy. Hello gulp pre-processing.