Color code last file date based on how long ago it was published
Last file date colors are the same as for score:

* Red means over a year ago
* Yellow means less than a year ago but more than 3 months ago
* Green means within the last three months

This is not 100% accurate, for example we don't go out of our way
to work exactly based on current date, this works from beginning of
month. So if today is 29th of August. Green will be anything that
was published between 1st of May and now. If it's 2nd of August
today, green will still be between 1st of May and now, so we're
looking at months, not dates.